Almost completely ditch src/util/clvm.py.
